Validação de usuario, laboratório e categorias. Relatórios funcionando. Deletando Categorias e Laboratorios.

+++ database/migrations/2020_08_06_065756_create_suggestions_table.php
@@ -15,7 +15,7 @@ class CreateSuggestionsTable extends Migration
     {
         Schema::create('suggestions', function (Blueprint $table) {
             $table->increments('id');
-			$table->integer('user_registration')->unsigned();
+			$table->bigInteger('user_registration')->unsigned();
 			$table->text('description');
 
             $table->foreign('user_registration')->references('registration')->on('users');
